+++ plotnine/stats/__init__.py
@@ -13,6 +13,7 @@ from .stat_identity import stat_identity
 from .stat_qq import stat_qq
 from .stat_qq_line import stat_qq_line
 from .stat_quantile import stat_quantile
+from .stat_sina import stat_sina
 from .stat_smooth import stat_smooth
 from .stat_sum import stat_sum
 from .stat_summary import stat_summary
@@ -30,6 +31,7 @@ __all__ = ['stat_count', 'stat_bin',
            'stat_identity',
            'stat_qq', 'stat_qq_line',
            'stat_quantile',
+           'stat_sina',
            'stat_smooth', 'stat_sum',
            'stat_summary', 'stat_summary_bin',
            'stat_unique', 'stat_ydensity']